refactor: 복잡한 로직 함수 추출 (#72)
